fix(accessibility): ensure newValue is a double in slider semantics

+++ lib/shared/accessibility/mixins/accessibility_mixins.dart
@@ -116,13 +116,17 @@ mixin SemanticWrapperMixin {
       hint: hint,
       onIncrease: onChanged != null
           ? () {
-              final newValue = (value + ((max - min) / 10)).clamp(min, max);
+              final newValue = (value + ((max - min) / 10))
+                  .clamp(min, max)
+                  .toDouble();
               onChanged(newValue);
             }
           : null,
       onDecrease: onChanged != null
           ? () {
-              final newValue = (value - ((max - min) / 10)).clamp(min, max);
+              final newValue = (value - ((max - min) / 10))
+                  .clamp(min, max)
+                  .toDouble();
               onChanged(newValue);
             }
           : null,
